metrics: remove "name" from the "unmatched_count" metric
In the standup this morning, we discussed the problem with the label
"name": it may end up being very big (cardinality problem) in case of
many CRs being created at the same time and if the metrics are scrapped
right before the approver has had time to set the Approved condition.

Also, it is clear that the CR "name" isn't often useful; it would be
much more useful to have a label "owner_name" and "owner_kind".

Signed-off-by: Maël Valais <mael@vls.dev>